move test config files to tests dir
