+++ makefile
@@ -1,11 +1,11 @@
 mppsolar-tests: 
-	python3 -m unittest discover -s tests/mppsolar_tests -f 
+	python3 -m unittest discover -s tests/mpp -f 
 	
 powermon-unit-tests: 
-	python3 -m unittest discover -s tests/powermon_tests/unit -f -v
+	python3 -m unittest discover -s tests/pmon/unit -f -v
 
 powermon-integration-tests: 
-	python3 -m unittest discover -s tests/powermon_tests/integration -f -v
+	python3 -m unittest discover -s tests/pmon/integration -f -v
 
 test:
 	python3 -m unittest discover -s tests -f
