+++ tests/mpp/test_cmd_mppsolar.py
