+++ tests/mpp/test_cmd_mppsolar_pi30revo.py
