+++ tests/mpp/test_output_mqtt.py
