+++ tests/pmon/config/bad.yaml
