+++ tests/pmon/config/min.yaml
@@ -1,0 +1,41 @@
+# minimal config for powermon
+device:
+  name: Test_Inverter
+  id: A123456789
+  model: 8048MAX
+  manufacturer: MPP-Solar
+  port:
+    type: test
+    # path: /dev/ttyUSB0
+    # response_number: 1
+    protocol: PI30MAX
+
+commands:
+- command: QEY2021
+  outputs:
+  - type: screen
+    format: 
+      type: table
+      draw_lines: True
+      extra_info: True
+# - command: QMCHGCR
+#   outputs:
+#   - type: screen
+#     format: 
+#       type: table
+#       draw_lines: True
+#       extra_info: True
+# - command: F50
+#   override:
+#     temperature: F
+#     pv1_charging_power:
+#       units: kW
+#   trigger:
+#     every: 5  # every 5 seconds
+#   outputs:
+#   - type: screen
+#     format:
+#       type: table
+#       draw_lines: True
+#       extra_info: True
+loop: once
