+++ tests/pmon/config/powermon-hass.yaml
