+++ tests/pmon/config/powermon-qed.yaml
