+++ tests/pmon/config/powermon.yaml
