+++ tests/pmon/integration/test_cmd_powermon_pi30.py
