+++ tests/pmon/integration/test_cmd_powermon_pi30max.py
