+++ tests/pmon/integration/test_config_validation.py
