+++ tests/pmon/unit/test_command_definition.py
