+++ tests/pmon/unit/test_powermon_output_mqtt.py
